moving validation for UserController methods to middlewares to be called in the router
